Started work on creating training set
Added a python notebook that reads in the JSON file (which is not
currently in the repo — not sure if it should be?) and loads some of
the data into a pandas data frame. For now just playing around with how
to automatically tag some of the tweets as trolls by identifying bad
users (who use bad words). This obviously won’t work perfectly so need
to look at what that is missing.